(m2): upload proof of practice: aligned the page with the hi-fidelity prototype
